clipflow-api/pyproject.toml
[tool.poetry]
name = "clipflow-api"
version = "0.1.0"
description = "Production backend API for ClipFlow"
authors = ["ClipFlow"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.6", extras = ["standard"] }
sqlalchemy = "^2.0.35"
psycopg = { version = "^3.2.1", extras = ["binary"] }
alembic = "^1.13.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
PyJWT = "^2.9.0"
python-multipart = "^0.0.12"
redis = "^5.1.1"
minio = "^7.2.8"
httpx = "^0.27.2"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.8"
mypy = "^1.11.2"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"